requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
cffi==2.0.0
click==8.1.8
cryptography==46.0.1
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-utils==0.8.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
mypy-extensions==1.1.0
passlib==1.7.4
pillow==11.3.0
psutil==5.9.8
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
starlette==0.47.3
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
uvicorn==0.35.0
